requirements.txt
# Bot Leboncoin - Dépendances Python
# Compatible Python 3.8+

# Requêtes HTTP avec support async
httpx>=0.25.0

# Parsing HTML
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Bot Telegram
python-telegram-bot>=20.7

# Base de données
sqlalchemy>=2.0.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Utilitaires (optionnels mais recommandés)
# Pour typage avancé si Python < 3.9
typing-extensions>=4.0.0; python_version < "3.9"
